Clean up old files and add new drawer components
